divided endpoint codebase
