+++ main.go
@@ -2,14 +2,12 @@ package main
 
 import (
 	"database/sql"
-	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
 	"net/http"
 	"os"
 	"strconv"
-	"strings"
 	"time"
 
 	"golang.org/x/net/context"
@@ -111,7 +109,7 @@ func main() {
 		var subject string
 
 		err := authenticate(request, &subject)
-		
+
 		if err != nil {
 			writer.WriteHeader(401)
 			fmt.Fprintln(writer, err.Error())
@@ -166,285 +164,6 @@ func stamp_message(channel_id int64, message *Message) error {
 	}
 
 	message.Id = id
-
-	return nil
-}
-
-func endpoint_friendships(writer http.ResponseWriter, request *http.Request) *http_status {
-	var subject string
-
-	err := authenticate(request, &subject)
-	
-	if err!= nil {
-		return &http_status{401, err.Error()}
-	}
-
-	person_id, err := strconv.Atoi(strings.TrimPrefix(request.URL.Path, "/friendships/"))
-
-	if request.Method != "GET" && err != nil {
-		return &http_status{400, "failed to parse person id"}
-	}
-
-	switch request.Method {
-	case "GET":
-
-	case "PUT":
-		if _, err := db.Exec("INSERT INTO friendships (person_0, person_1) VALUES (?, ?)", subject, person_id); err != nil {
-			return &http_status{500, err.Error()}
-		}
-	case "DELETE":
-		if _, err := db.Exec("DELETE FROM friendships WHERE person_0 = ? AND person_1 = ?", subject, person_id); err != nil {
-			return &http_status{500, err.Error()}
-		}
-	default:
-		return &http_status{405, "method not allowed"}
-	}
-
-	rows, err := db.Query("SELECT person_1 FROM friendships WHERE person_0 = ? AND person_1 = id", subject)
-
-	if err != nil {
-		return &http_status{500, err.Error()}
-	}
-
-	friends := make([]string, 0, 16)
-
-	var person string
-
-	for rows.Next() {
-		if err := rows.Scan(&person); err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		friends = append(friends, person)
-	}
-
-	buffer, err := json.Marshal(friends)
-
-	if err != nil {
-		return &http_status{500, err.Error()}
-	}
-
-	writer.Header().Set("Content-Type", "application/json")
-	writer.Write(buffer)
-
-	return nil
-}
-
-func endpoint_channels(writer http.ResponseWriter, request *http.Request) *http_status {
-	var subject string
-
-	err := authenticate(request, &subject)
-	
-	if err != nil {
-		return &http_status{401, err.Error()}
-	}
-
-	if err := request.ParseForm(); err != nil {
-		return &http_status{401, "auth failed"}
-	}
-
-	parameter := strings.TrimPrefix(request.URL.Path, "/channels/")
-
-	if request.Method == "GET" && parameter == "" {
-		rows, err := db.Query("SELECT id, name FROM participations, channels WHERE person = ? AND channel = id", subject)
-
-		var channel Channel
-
-		channels := make([]Channel, 0, 16)
-
-		for rows.Next() {
-			if err := rows.Scan(&channel.Id, &channel.Name); err != nil {
-				return &http_status{500, err.Error()}
-			}
-
-			channels = append(channels, channel)
-		}
-
-		buffer, err := json.Marshal(channels)
-
-		if err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		writer.Header().Set("Content-Type", "application/json")
-		writer.Write(buffer)
-
-		return nil
-	}
-
-	if request.Method == "POST" && parameter == "" {
-		tx, err := db.Begin()
-
-		if err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		result, err := tx.Exec("INSERT INTO channels (name) VALUES (?)", request.PostForm.Get("name"))
-
-		if err != nil {
-			tx.Rollback()
-			return &http_status{500, err.Error()}
-		}
-
-		channel_id, err := result.LastInsertId()
-
-		if err != nil {
-			tx.Rollback()
-			return &http_status{500, err.Error()}
-		}
-
-		if _, err = tx.Exec("INSERT INTO participations (person, channel) VALUES (?, ?)", subject, channel_id); err != nil {
-			tx.Rollback()
-			return &http_status{500, err.Error()}
-		}
-
-		if err := tx.Commit(); err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		fmt.Fprintf(writer, "%d", channel_id)
-
-		return nil
-	}
-
-	channel_id, err := strconv.Atoi(parameter)
-
-	if err != nil {
-		return &http_status{400, "failed to parse channel id"}
-	}
-
-	switch request.Method {
-	case "GET":
-
-	case "POST":
-		if _, err := db.Exec("INSERT INTO participations (channel, person) VALUES (?, ?)", channel_id, subject); err != nil {
-			return &http_status{500, err.Error()}
-		}
-	case "DELETE":
-		if _, err := db.Exec("DELETE FROM participations WHERE channel = ? AND person = ?", channel_id, subject); err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		tx, err := db.Begin()
-
-		if err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		row := tx.QueryRow("SELECT COUNT(*) FROM participations WHERE channel = ?", channel_id)
-
-		var size int
-
-		if err := row.Scan(&size); err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		if size == 0 {
-			if _, err := tx.Exec("DELETE FROM channels WHERE id = ?", channel_id); err != nil {
-				return &http_status{500, err.Error()}
-			}
-		}
-
-		tx.Commit()
-	default:
-		return &http_status{405, "method not allowed"}
-	}
-
-	row := db.QueryRow("SELECT name FROM channels WHERE id = ?", channel_id)
-
-	if err != nil {
-		return &http_status{500, err.Error()}
-	}
-
-	var channel struct {
-		Name         string
-		Participants []string
-	}
-
-	if err := row.Scan(&channel.Name); err != nil {
-		return &http_status{410, err.Error()}
-	}
-
-	rows, err := db.Query("SELECT person FROM participations WHERE channel = ?", channel_id)
-	channel.Participants = make([]string, 0, 16)
-
-	if err != nil {
-		return &http_status{500, err.Error()}
-	}
-
-	var person string
-
-	for rows.Next() {
-		if err := rows.Scan(&person); err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		channel.Participants = append(channel.Participants, person)
-	}
-
-	buffer, err := json.Marshal(channel)
-
-	if err != nil {
-		return &http_status{500, err.Error()}
-	}
-
-	writer.Header().Set("Content-Type", "application/json")
-	writer.Write(buffer)
-
-	return nil
-}
-
-func endpoint_messages(writer http.ResponseWriter, request *http.Request) *http_status {
-	var subject string
-
-	err := authenticate(request, &subject)
-	
-	if err != nil {
-		return &http_status{401, err.Error()}
-	}
-
-	if request.Method != "GET" {
-		return &http_status{405, "method not allowed"}
-	}
-
-	channel_id, err := strconv.Atoi(strings.TrimPrefix(request.URL.Path, "/messages/"))
-
-	if err != nil {
-		return &http_status{400, "invalid channel name"}
-	}
-
-	since_id, err := strconv.Atoi(request.FormValue("since_id"))
-
-	if err != nil {
-		return &http_status{400, "invalid since_id"}
-	}
-
-	rows, err := db.Query("SELECT id, channel, author, is_event, posted_at, content FROM messages WHERE channel = ? AND id > ?", channel_id, since_id)
-
-	if err != nil {
-		return &http_status{500, err.Error()}
-	}
-
-	defer rows.Close()
-
-	var message Message
-
-	for rows.Next() {
-		err := rows.Scan(&message.Id, &message.Channel, &message.Author, &message.IsEvent, &message.PostedAt, &message.Content)
-
-		if err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		buffer, err := json.Marshal(message)
-
-		if err != nil {
-			return &http_status{500, err.Error()}
-		}
-
-		writer.Header().Set("Content-Type", "application/json")
-		writer.Write(buffer)
-	}
 
 	return nil
 }
